+++ C_code_vitis/trees_managment.h
@@ -14,7 +14,7 @@
 
 
 void generate_rando_trees(tree_data trees[N_TREES][N_NODE_AND_LEAFS], 
-                    uint8_t n_features, uint8_t n_trees, float max_features[N_FEATURE], float min_features[N_FEATURE]);
+                    uint8_t n_features, uint16_t n_trees, float max_features[N_FEATURE], float min_features[N_FEATURE]);
 
 void mutate_population(tree_data trees_population[POPULATION][N_TREES][N_NODE_AND_LEAFS],
                         float population_accuracy[POPULATION], float max_features[N_FEATURE],
